--- src/Components/Cards/index.js
@@ -1,3 +1,0 @@
-export { ProductCard } from "./ProductCard";
-export { CartCard } from "./CartCard";
-export { WishListCard } from "./WishListCard";
