+++ src/Components/WishListCard.js
@@ -1,8 +1,8 @@
-import { useData } from "../../Contexts";
-import "../../styles.css";
-import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
+import { useData } from "../Contexts";
+import "../styles.css";
+import { Link } from "react-router-dom";
 
-export function ProductCard({product}) {
+export function WishListCard({product}) {
     let {
         id,
         name,
@@ -25,21 +25,7 @@ export function ProductCard({product}) {
         }
         return(
             <div>
-                <button class="btn btn-primary" onClick={() => dispatch({type: "MANIPULATE_CART", payload: {manipulation: "ADD_TO_CART", item: product}})} >Add To Cart</button>
-            </div>
-        )
-    }
-    const WishListButtons = () => {
-        if(itemsInWishList.find((item) => item.id === id)){
-            return(
-                <div>
-                    <button class="btn btn-tertiary" onClick={() => dispatch({type: "MANIPULATE_WISHLIST", payload: {manipulation: "REMOVE_FROM_WISHLIST", item: product}})} >Remove from Wishlist</button>
-                </div>
-            )
-        }
-        return(
-            <div>
-                <button class="btn btn-tertiary" onClick={() => dispatch({type: "MANIPULATE_WISHLIST", payload: {manipulation: "ADD_TO_WISHLIST", item: product}})} >Add to WishList</button>
+                <button class="btn btn-primary" onClick={() => dispatch({type: "MANIPULATE_CART", payload: {manipulation: "MOVE_TO_CART", item: product}})} >Move To Cart</button>
             </div>
         )
     }
@@ -56,7 +42,9 @@ export function ProductCard({product}) {
                 {inStock && <div> In Stock </div>}
                 {!inStock && <div> Out of Stock </div>}
                 {fastDelivery ? ( <div> Fast Delivery </div> ) : ( <div> 3 days minimum </div> )}
-                {WishListButtons()}
+                <div>
+                    <button class="btn btn-tertiary" onClick={() => dispatch({type: "MANIPULATE_WISHLIST", payload: {manipulation: "REMOVE_FROM_WISHLIST", item: product}})} >Remove from Wishlist</button>
+                </div>
                 {CartButtons()}
             </div>
         </div>
