+++ src/Components/index.js
@@ -1,6 +1,3 @@
-export { Cart } from "./Cart";
-export { ProductListing } from "./ProductListing";
-export { WishList } from "./WishList";
-export { Login } from "./Login";
-export { Signup } from "./Signup";
-export { Logout } from "./Logout";
+export { ProductCard } from "./ProductCard";
+export { CartCard } from "./CartCard";
+export { WishListCard } from "./WishListCard";
