+++ src/Pages/index.js
@@ -1,0 +1,6 @@
+export { Cart } from "./Cart";
+export { ProductListing } from "./ProductListing";
+export { WishList } from "./WishList";
+export { Login } from "./Login";
+export { Signup } from "./Signup";
+export { Logout } from "./Logout";
